+++ src/views/MonthView.tsx
@@ -12,22 +12,24 @@ function getMonthTitle(year: number, month: number) {
 interface IProps {
   year: number;
   month: number;
-  selectedDates: Set<number>;
-  referenceDate: number;
-  setReferenceDate: (datetime: number) => void;
 }
 
-const MonthView: React.FC<IProps> = (props) => {
-  const { year, month } = props;
+const MonthView: React.FC<IProps> = ({ year, month }) => {
   const title = getMonthTitle(year, month);
   const offset = new Date(Date.UTC(year, month, 1)).getISODay() - 1;
+
   return (
     <View style={styles.container}>
       <Text style={styles.text}>{title}</Text>
       {Array.from({ length: 6 }, (_, i) => (
         <View key={i} style={styles.row}>
           {Array.from({ length: 7 }, (_, j) => (
-            <DateView key={j} day={1 - offset + 7 * i + j} {...props} />
+            <DateView
+              key={j}
+              year={year}
+              month={month}
+              day={1 - offset + 7 * i + j}
+            />
           ))}
         </View>
       ))}
